+++ TourPlanner.DAL/Common/IFileAccess.cs
@@ -8,8 +8,8 @@ namespace TourPlanner.DAL.Common
 {
     public interface IFileAccess
     {
-        int CreateNewTourFile(string name, string description, string from, string to, string routeInformation, double distance);
-        int CreateNewLogFile(Tour tourLogItem, string date, string totalTime, string report, double distance, int rating,
+        int CreateNewTourFile(string name, string description, string from, string to, string routeInformation, int distance);
+        int CreateNewLogFile(Tour tourLogItem, string date, string totalTime, string report, int distance, int rating,
             int averageSpeed, int maxSpeed, int minSpeed, int averageStepCount, int burntCalories);
         IEnumerable<FileInfo> SearchFiles(string searchTerm, MediaType searchType);
