+++ TourPlanner.DAL/DAO/ILogDAO.cs
@@ -8,7 +8,7 @@ namespace TourPlanner.DAL.DAO
     public interface ILogDAO
     {
         TourLog FindById(int id);
-        TourLog AddNewItemLog(Tour tourLogItem, string date, string totalTime, string report, double distance, int rating,
+        TourLog AddNewItemLog(Tour tourLogItem, string date, string totalTime, string report, int distance, int rating,
             int averageSpeed, int maxSpeed, int minSpeed, int averageStepCount, int burntCalories);
         IEnumerable<TourLog> GetLogsForTour(Tour tourItem);
     }
